[fel]fix write32

+++ main.c
@@ -241,7 +241,7 @@ int main(int argc, char * argv[])
 		if(argc == 2)
 		{
 			uint32_t addr = strtoul(argv[0], NULL, 0);
-			size_t val = strtoul(argv[1], NULL, 0);
+			uint32_t val = strtoul(argv[1], NULL, 0);
 			fel_write32(&ctx, addr, val);
 		}
 		else
